+++ packages/playground/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @talend/ui-playground
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [df182bbd4]
+  - @talend/react-cmf-router@5.0.0
+
 ## 1.0.5
 
 ### Patch Changes